[FE] fix: 레이아웃 수정사항 반영
