+++ front/src/components/Header.js
@@ -1,219 +1,105 @@
-import { useState } from 'react';
-import AppBar from '@mui/material/AppBar';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import IconButton from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
-import Menu from '@mui/material/Menu';
-import MenuIcon from '@mui/icons-material/Menu';
-import Container from '@mui/material/Container';
-import Avatar from '@mui/material/Avatar';
-import Button from '@mui/material/Button';
-import Tooltip from '@mui/material/Tooltip';
-import MenuItem from '@mui/material/MenuItem';
-import AdbIcon from '@mui/icons-material/Adb';
-import SearchIcon from '@mui/icons-material/Search';
-import InputBase from '@mui/material/InputBase';
-import { styled as muiStyled, alpha } from '@mui/material/styles';
-import styled from 'styled-components';
+import styled, { ThemeProvider } from 'styled-components';
+import theme from '../styles/theme';
+import { Link } from 'react-router-dom';
+import Searchbar from './Searchbar';
 
-const pages = ['컬렉션', '페어링'];
-const settings = ['Profile', 'Account', 'Dashboard', 'Logout'];
-
-const Search = muiStyled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginRight: theme.spacing(2),
-  marginLeft: 0,
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing(3),
-    width: 'auto',
-  },
-}));
-
-const SearchIconWrapper = muiStyled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-const StyledInputBase = muiStyled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('md')]: {
-      width: '20ch',
-    },
-  },
-}));
-
-const Version = styled.span`
-  display: none;
+const HeaderContainer = styled.div`
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 60px;
+  background-color: white;
+  border-bottom: 1px solid ${({ theme }) => theme.colors.lightgray};
+  &:hover {
+    cursor: pointer;
+  }
+  display: flex;
+  align-items: center;
 `;
 
-const Header = () => {
-  const [anchorElNav, setAnchorElNav] = useState(null);
-  const [anchorElUser, setAnchorElUser] = useState(null);
+const LogoContainer = styled.div`
+  margin: 0 40px;
+`;
 
-  const handleOpenNavMenu = (event) => {
-    setAnchorElNav(event.currentTarget);
-  };
-  const handleOpenUserMenu = (event) => {
-    setAnchorElUser(event.currentTarget);
-  };
+const HeaderBtn = styled.button`
+  background-color: transparent;
+  font-size: 16px;
+  font-weight: 700;
+  border: none;
+  margin: 0 15px;
+  width: 60px;
+  &:hover {
+    cursor: pointer;
+  }
+`;
 
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
-  };
+const CollectionBtn = styled(HeaderBtn)``;
+const PairingBtn = styled(HeaderBtn)`
+  margin-right: 330px;
+`;
 
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
-  };
+const LoginOutBtn = styled(HeaderBtn)`
+  font-size: 13px;
+  font-weight: 400;
+  &:hover {
+    cursor: pointer;
+  }
+  position: absolute;
+  top: 22px;
+  right: 80px;
+`;
 
+const MyPageIconContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  &:hover {
+    cursor: pointer;
+  }
+  img {
+    position: absolute;
+    right: 40px;
+  }
+`;
+
+const Header = ({ isLogin = false }) => {
   return (
-    <AppBar position="static">
-      <Version>version: {process.env.REACT_APP_SERVICE_VERSION}</Version>
-      <Container maxWidth="xl">
-        <Toolbar disableGutters>
-          <AdbIcon sx={{ display: { xs: 'none', md: 'flex' }, mr: 1 }} />
-          <Typography
-            variant="h6"
-            noWrap
-            component="a"
-            href="/"
-            sx={{
-              mr: 2,
-              display: { xs: 'none', md: 'flex' },
-              fontFamily: 'monospace',
-              fontWeight: 700,
-              letterSpacing: '.3rem',
-              color: 'inherit',
-              textDecoration: 'none',
-            }}
-          >
-            LOGO
-          </Typography>
-
-          <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
-            <IconButton
-              size="large"
-              aria-label="account of current user"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenNavMenu}
-              color="inherit"
-            >
-              <MenuIcon />
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              anchorEl={anchorElNav}
-              anchorOrigin={{
-                vertical: 'bottom',
-                horizontal: 'left',
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: 'top',
-                horizontal: 'left',
-              }}
-              open={Boolean(anchorElNav)}
-              onClose={handleCloseNavMenu}
-              sx={{
-                display: { xs: 'block', md: 'none' },
-              }}
-            >
-              {pages.map((page) => (
-                <MenuItem key={page} onClick={handleCloseNavMenu}>
-                  <Typography textAlign="center">{page}</Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
-          <AdbIcon sx={{ display: { xs: 'flex', md: 'none' }, mr: 1 }} />
-          <Typography
-            variant="h5"
-            noWrap
-            component="a"
-            href=""
-            sx={{
-              mr: 2,
-              display: { xs: 'flex', md: 'none' },
-              flexGrow: 1,
-              fontFamily: 'monospace',
-              fontWeight: 700,
-              letterSpacing: '.3rem',
-              color: 'inherit',
-              textDecoration: 'none',
-            }}
-          >
-            LOGO
-          </Typography>
-          <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
-            {pages.map((page) => (
-              <Button
-                key={page}
-                onClick={handleCloseNavMenu}
-                sx={{ my: 2, color: 'white', display: 'block' }}
-              >
-                {page}
-              </Button>
-            ))}
-          </Box>
-          <Search>
-            <SearchIconWrapper>
-              <SearchIcon />
-            </SearchIconWrapper>
-            <StyledInputBase
-              placeholder="Search…"
-              inputProps={{ 'aria-label': 'search' }}
+    <ThemeProvider theme={theme}>
+      <HeaderContainer>
+        <LogoContainer>
+          <Link to="/">
+            <img
+              src={process.env.PUBLIC_URL + '/images/CherryPick.svg'}
+              alt="CherryPick main logo"
             />
-          </Search>
-          <Box sx={{ flexGrow: 0 }}>
-            <Tooltip title="Open settings">
-              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
-              </IconButton>
-            </Tooltip>
-            <Menu
-              sx={{ mt: '45px' }}
-              id="menu-appbar"
-              anchorEl={anchorElUser}
-              anchorOrigin={{
-                vertical: 'top',
-                horizontal: 'right',
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: 'top',
-                horizontal: 'right',
-              }}
-              open={Boolean(anchorElUser)}
-              onClose={handleCloseUserMenu}
-            >
-              {settings.map((setting) => (
-                <MenuItem key={setting} onClick={handleCloseUserMenu}>
-                  <Typography textAlign="center">{setting}</Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
-        </Toolbar>
-      </Container>
-    </AppBar>
+          </Link>
+        </LogoContainer>
+        <Link to="/collection">
+          <CollectionBtn>컬렉션</CollectionBtn>
+        </Link>
+        <Link to="/pairing">
+          <PairingBtn>페어링</PairingBtn>
+        </Link>
+        <Searchbar />
+        {isLogin ? (
+          <LoginOutBtn>로그아웃</LoginOutBtn>
+        ) : (
+          <Link to="/user/signin">
+            <LoginOutBtn>로그인</LoginOutBtn>
+          </Link>
+        )}
+        <Link to="/mypage">
+          <MyPageIconContainer>
+            <img
+              src={process.env.PUBLIC_URL + '/images/Mypage_Icon.svg'}
+              alt="Mypage Icon"
+            />
+          </MyPageIconContainer>
+        </Link>
+      </HeaderContainer>
+    </ThemeProvider>
   );
 };
+
 export default Header;
