+++ front/src/components/PageContainer.js
@@ -1,9 +1,8 @@
 import CssBaseline from '@mui/material/CssBaseline';
 import Box from '@mui/material/Box';
 import Container from '@mui/material/Container';
-import StickyFooter from './StickyFooter';
-//import Header from './Header';
-import Header from './Header_';
+import Header from './Header';
+import Footer from './Footer';
 
 const PageContainer = ({ children, header, footer }) => {
   return (
@@ -19,7 +18,7 @@ const PageContainer = ({ children, header, footer }) => {
       <Container component="main" sx={{ mt: 8, mb: 2 }} maxWidth="sm">
         {children}
       </Container>
-      {footer ? <StickyFooter /> : null}
+      {footer ? <Footer /> : null}
     </Box>
   );
 };
