--- front/src/components/StickyFooter.js
@@ -1,38 +1,0 @@
-import Box from '@mui/material/Box';
-import Typography from '@mui/material/Typography';
-import Container from '@mui/material/Container';
-
-const Copyright = () => {
-  return (
-    <Typography variant="body2" color="text.secondary">
-      {'Copyright © '}
-      Running_Potatoes {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-};
-
-const StickyFooter = () => {
-  return (
-    <Box
-      component="footer"
-      sx={{
-        py: 3,
-        px: 2,
-        mt: 'auto',
-        backgroundColor: (theme) =>
-          theme.palette.mode === 'light'
-            ? theme.palette.grey[200]
-            : theme.palette.grey[800],
-      }}
-    >
-      <Container maxWidth="sm">
-        <Typography variant="body1">
-          My sticky footer can be found here.
-        </Typography>
-        <Copyright />
-      </Container>
-    </Box>
-  );
-};
-export default StickyFooter;
